+++ tests/test_models.py
@@ -3,7 +3,8 @@ from decimal import Decimal
 
 import pytest
 from django.core.exceptions import ValidationError
-from django_prices import forms, widgets
+from django.db.models.fields import DecimalField
+from django_prices import forms, widgets, models
 from prices import Money, TaxedMoney
 
 from .models import Model, NullModel
@@ -73,3 +74,101 @@ def test_money_field_formfield_returns_form_with_select_input_if_choices_are_def
     form_field = field.formfield()
     assert isinstance(form_field, forms.MoneyField)
     assert isinstance(form_field.widget, widgets.MoneyInput)
+
+
+def test_money_field_init():
+    field = models.MoneyField(amount_field="amount", currency_field="currency")
+    assert field.get_default() == Money(0, None)
+    assert field.amount_field == "amount"
+    assert field.currency_field == "currency"
+
+
+def test_compare_money_field_with_same_type_field():
+    field_1 = models.MoneyField(amount_field="money_net_amount", currency_field="currency")
+    field_2 = models.MoneyField(amount_field="money_net_amount", currency_field="currency")
+
+    # Comparision is based on creation_counter attribute
+    assert field_1 < field_2
+    field_2.creation_counter -= 1
+    assert field_1 == field_2
+
+
+def test_compare_money_field_with_django_field():
+    field_1 = models.MoneyField(amount_field="money_net_amount", currency_field="currency")
+    field_2 = DecimalField(default="5", max_digits=9, decimal_places=2)
+
+    # Comparision is based on creation_counter attribute
+    assert field_1 < field_2
+    assert not field_1 > field_2
+    field_2.creation_counter -= 1
+    assert field_1 == field_2
+
+
+def test_compare_money_field_with_taxed_money_field():
+    field_1 = models.MoneyField(amount_field="money_net_amount", currency_field="currency")
+    field_2 = models.TaxedMoneyField(net_field="price_net", gross_field="price_gross")
+
+    # Comparision is based on creation_counter attribute
+    assert field_1 < field_2
+    assert not field_1 > field_2
+    field_2.creation_counter -= 1
+    assert field_1 == field_2
+
+
+def test_taxed_money_field_init():
+    field = models.TaxedMoneyField(
+        net_amount_field="price_net",
+        gross_amount_field="price_gross",
+        currency="currency",
+    )
+    assert field.net_amount_field == "price_net"
+    assert field.gross_amount_field == "price_gross"
+    assert field.currency == "currency"
+
+
+def test_compare_taxed_money_field_with_same_type_field():
+    field_1 = models.TaxedMoneyField(
+        net_amount_field="price_net",
+        gross_amount_field="price_gross",
+        currency="currency",
+    )
+    field_2 = models.TaxedMoneyField(
+        net_amount_field="price_net",
+        gross_amount_field="price_gross",
+        currency="currency",
+    )
+
+    # Comparision is based on creation_counter attribute
+    assert field_1 < field_2
+    field_2.creation_counter -= 1
+    assert field_1 == field_2
+
+
+def test_compare_taxed_money_field_with_django_field():
+    field_1 = models.TaxedMoneyField(
+        net_amount_field="price_net",
+        gross_amount_field="price_gross",
+        currency="currency",
+    )
+    field_2 = DecimalField(default="5", max_digits=9, decimal_places=2)
+
+    # Comparision is based on creation_counter attribute
+    assert field_1 < field_2
+    assert not field_1 > field_2
+    field_2.creation_counter -= 1
+    assert field_1 == field_2
+
+
+def test_compare_taxed_money_field_with_money_field():
+    field_1 = models.TaxedMoneyField(
+        net_amount_field="price_net",
+        gross_amount_field="price_gross",
+        currency="currency",
+    )
+    field_2 = models.MoneyField(amount_field="money_net_amount", currency_field="currency")
+
+    # Comparision is based on creation_counter attribute
+    assert field_1 < field_2
+    assert not field_1 > field_2
+    field_2.creation_counter -= 1
+    assert field_1 == field_2
